+++ lib/ansible/cli/config.py
@@ -36,28 +36,30 @@ class ConfigCLI(CLI):
 
     def init_parser(self):
 
-        self.parser = super(ConfigCLI, self).init_parser(
+        super(ConfigCLI, self).init_parser(
             usage="usage: %%prog [%s] [--help] [options] [ansible.cfg]" % "|".join(sorted(self.VALID_ACTIONS)),
             epilog="\nSee '%s <command> --help' for more information on a specific command.\n\n" % os.path.basename(sys.argv[0]),
             desc="View, edit, and manage ansible configuration.",
         )
-        self.parser.add_option('-c', '--config', dest='config_file', help="path to configuration file, defaults to first file found in precedence.")
+        self.parser.add_option('-c', '--config', dest='config_file',
+                               help="path to configuration file, defaults to first file found in precedence.")
 
         self.set_action()
 
         # options specific to self.actions
         if self.action == "list":
             self.parser.set_usage("usage: %prog list [options] ")
-        if self.action == "dump":
+
+        elif self.action == "dump":
             self.parser.add_option('--only-changed', dest='only_changed', action='store_true',
                                    help="Only show configurations that have changed from the default")
+
         elif self.action == "update":
             self.parser.add_option('-s', '--setting', dest='setting', help="config setting, the section defaults to 'defaults'")
             self.parser.set_usage("usage: %prog update [options] [-c ansible.cfg] -s '[section.]setting=value'")
+
         elif self.action == "search":
             self.parser.set_usage("usage: %prog update [options] [-c ansible.cfg] <search term>")
-
-        return self.parser
 
     def post_process_args(self, options, args):
         super(ConfigCLI, self).post_process_args(options, args)
