+++ test/units/cli/test_cli.py
@@ -31,11 +31,6 @@ from ansible import cli
 
 class TestCliVersion(unittest.TestCase):
 
-    def test_version(self):
-        ver = cli.CLI.version('ansible-cli-test')
-        self.assertIn('ansible-cli-test', ver)
-        self.assertIn('python version', ver)
-
     def test_version_info(self):
         version_info = cli.CLI.version_info()
         self.assertEqual(version_info['string'], __version__)
