+++ test/units/playbook/test_play_context.py
@@ -11,10 +11,10 @@ import os
 
 import pytest
 
-from ansible import arguments
 from ansible import constants as C
 from ansible import context
-from ansible import cli
+from ansible.arguments import context_objects as co
+from ansible.arguments import optparse_helpers as opt_help
 from units.compat import unittest
 from ansible.errors import AnsibleError, AnsibleParserError
 from ansible.module_utils.six.moves import shlex_quote
@@ -25,19 +25,26 @@ from units.mock.loader import DictDataLoader
 
 @pytest.fixture
 def parser():
-    parser = cli.base_parser(runas_opts=True, meta_opts=True,
-                             runtask_opts=True, vault_opts=True,
-                             async_opts=True, connect_opts=True,
-                             subset_opts=True, check_opts=True,
-                             inventory_opts=True,)
+    parser = opt_help.create_base_parser()
+
+    opt_help.add_runas_options(parser)
+    opt_help.add_meta_options(parser)
+    opt_help.add_runtask_options(parser)
+    opt_help.add_vault_options(parser)
+    opt_help.add_async_options(parser)
+    opt_help.add_connect_options(parser)
+    opt_help.add_subset_options(parser)
+    opt_help.add_check_options(parser)
+    opt_help.add_inventory_options(parser)
+
     return parser
 
 
 @pytest.fixture
 def reset_cli_args():
-    arguments.GlobalCLIArgs._Singleton__instance = None
+    co.GlobalCLIArgs._Singleton__instance = None
     yield
-    arguments.GlobalCLIArgs._Singleton__instance = None
+    co.GlobalCLIArgs._Singleton__instance = None
 
 
 def test_play_context(mocker, parser, reset_cli_args):
